requirements.txt
streamlit==1.30.0
streamlit-webrtc>=0.47.0
streamlit-drawable-canvas

numpy
pandas
scikit-learn
matplotlib
scipy
plotly

opencv-python-headless

google-generativeai
openai
python-dotenv

av

torch
torchvision

ultralytics

pillow>=10.4.0,<11
